Adicionando histórico de modificações em tarefas

+++ src/TaskManager.Infra.Data/Repositories/ProjectRepository.cs
@@ -34,6 +34,11 @@ public class ProjectRepository : IProjectRepository
         return await _context.Projects.Where(p => p.UsuarioId == userId).ToListAsync();
     }
 
+    public async Task<IEnumerable<Project>> GetByUserName(string userName)
+    {
+        return await _context.Projects.Where(p => p.Usuario.Nome == userName).ToListAsync();
+    }
+
     public async Task<IEnumerable<Project>> GetByDescricao(string descricao)
     {
         return await _context.Projects.Where(p => p.Descricao.Contains(descricao)).ToListAsync();
@@ -65,8 +70,17 @@ public class ProjectRepository : IProjectRepository
 
     public async Task<Project> Remove(Project project)
     {
+        if (!await CanRemoveProject(project.Id))
+            throw new Exception("Não é possível remover um projeto com tarefas não concluídas.");
+
         _context.Projects.Remove(project);
         await _context.SaveChangesAsync();
         return project;
+    }
+
+    public async Task<bool> CanRemoveProject(Guid id)
+    {
+        var project = await _context.Projects.Include(p => p.Tarefas).FirstOrDefaultAsync(p => p.Id == id);
+        return project != null && project.Tarefas.All(t => t.Concluida);
     }
 }
